refactor: Renamed analytics file
