+++ src/routes/ErrorPage.tsx
@@ -2,7 +2,7 @@ import { Button } from "antd";
 import React, { ReactElement } from "react";
 import { ErrorResponse, Link, useRouteError } from "react-router-dom";
 
-import { AnalyticsEvent, triggerAnalyticsEvent } from "../colorizer/utils/analytics_utils";
+import { AnalyticsEvent, triggerAnalyticsEvent } from "../colorizer/utils/analytics";
 import { FlexColumnAlignCenter } from "../styles/utils";
 
 import Header from "../components/Header";
